Time:   Fixed some VS warnings

+++ modules/tracktion_engine/playback/graph/tracktion_WaveNode.cpp
@@ -158,7 +158,7 @@ public:
                 readPos -= loopLength;
 
             startOffsetInDestBuffer += numToRead;
-            numSamplesToDo -= numToRead;
+            numSamplesToDo -= (int) numToRead;
         }
 
         return allOk;
